# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/projects/{project_id}/applications/{application_id}/releases:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
      ]
    get:
      operationId: listReleases
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses: { '200': { $ref: '#/components/responses/ReleasePage' } }
    post:
      operationId: createRelease
      requestBody: { $ref: '#/components/requestBodies/CreateRelease' }
      responses: { '201': { $ref: '#/components/responses/Release' } }
  /api/v1/projects/{project_id}/applications/{application_id}/releases/{release_id}:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
        { $ref: '#/components/parameters/ReleaseId' },
      ]
    get:
      operationId: getRelease
      responses: { '200': { $ref: '#/components/responses/Release' } }
  /api/v1/projects/{project_id}/applications/{application_id}/releases/{target_id}/runtime-diff:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
        { $ref: '#/components/parameters/TargetId' },
      ]
    get:
      operationId: getRuntimeDiff
      parameters:
        - { name: baseline_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses: { '200': { $ref: '#/components/responses/RuntimeDiff' } }
  /api/v1/projects/{project_id}/applications/{application_id}/releases/{target_id}/runtime-diff/summary:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
        { $ref: '#/components/parameters/TargetId' },
      ]
    get:
      operationId: getRuntimeDiffSummary
      description: Returns complete comparison classification totals and the largest absolute occurrence-count changes independent of Runtime Diff pagination.
      parameters:
        - { name: baseline_id, in: query, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 10, default: 5 } }
      responses:
        {
          '200': { $ref: '#/components/responses/RuntimeDiffSummary' },
          '400': { $ref: '#/components/responses/Error' },
          '401': { $ref: '#/components/responses/Error' },
          '404': { $ref: '#/components/responses/Error' },
        }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
    adminAuth: { type: http, scheme: bearer, description: System administrator credential }
  parameters:
    Cursor:
      {
        name: cursor,
        in: query,
        schema: { type: string, format: uuid },
        description: Opaque cursor scoped to the authenticated collection.,
      }
...
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    ProjectId:
      { name: project_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    ApplicationId:
      { name: application_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    ReleaseId:
      { name: release_id, in: path, required: true, schema: { type: string, format: uuid } }
    TargetId: { name: target_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message, request_id]
      properties:
        error: { type: string, example: not_found }
        message: { type: string, example: resource not found }
        request_id: { type: string, example: 0ec02ed2-8483-4981-893e-bffc535897d7 }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Present for field-level validation failures.
...
    RuntimeEventSemanticSummary:
      oneOf:
        - { $ref: '#/components/schemas/ProcessExecSemanticSummary' }
        - { $ref: '#/components/schemas/SyscallSemanticSummary' }
        - { $ref: '#/components/schemas/NetworkConnectSemanticSummary' }
        - { $ref: '#/components/schemas/InboundNetworkSemanticSummary' }
        - { $ref: '#/components/schemas/NetworkDnsQuerySemanticSummary' }
        - { $ref: '#/components/schemas/NetworkDnsResponseSemanticSummary' }
        - { $ref: '#/components/schemas/FileActivitySemanticSummary' }
        - { $ref: '#/components/schemas/ProcessExitSemanticSummary' }
        - { $ref: '#/components/schemas/ContainerTerminationSemanticSummary' }
        - { $ref: '#/components/schemas/ContainerRestartSemanticSummary' }
        - { $ref: '#/components/schemas/RestartLoopSemanticSummary' }
    ProcessExecSemanticSummary:
      type: object
      additionalProperties: false
      required: [executable]
      properties:
        executable: { type: string, minLength: 1 }
    SyscallSemanticSummary:
      type: object
      additionalProperties: false
      required: [process_command, syscall]
      properties:
        process_command: { type: string, minLength: 1 }
        syscall: { type: string, minLength: 1 }
    NetworkConnectSemanticSummary:
      type: object
      additionalProperties: false
      required: [process_command, address_family, destination_address, destination_port]
      properties:
        process_command: { type: string, minLength: 1, example: curl }
        address_family: { type: string, enum: [ipv4, ipv6] }
        destination_address: { type: string, format: ip, example: 203.0.113.7 }
        destination_port: { type: integer, minimum: 1, maximum: 65535, example: 443 }
        dns_context: { $ref: '#/components/schemas/DnsContext' }
    InboundNetworkSemanticSummary:
      type: object
      additionalProperties: false
      required: [process_command, transport, address_family, local_address, local_port]
      properties:
        process_command: { type: string, minLength: 1, example: payments }
        transport: { type: string, const: tcp }
        address_family: { type: string, enum: [ipv4, ipv6] }
        local_address: { type: string, format: ip, example: 0.0.0.0 }
        local_port: { type: integer, minimum: 1, maximum: 65535, example: 8080 }
    NetworkDnsQuerySemanticSummary:
      type: object
      additionalProperties: false
      required: [process_command, name, query_type, transport, direction]
      properties:
        process_command: { type: string, minLength: 1 }
        name: { $ref: '#/components/schemas/DnsName' }
        query_type: { $ref: '#/components/schemas/DnsQueryType' }
        transport: { $ref: '#/components/schemas/DnsTransport' }
        direction: { $ref: '#/components/schemas/DnsDirection' }
    NetworkDnsResponseSemanticSummary:
      type: object
      additionalProperties: false
      required: [process_command, name, query_type, response_code, transport, direction]
      properties:
        process_command: { type: string, minLength: 1 }
        name: { $ref: '#/components/schemas/DnsName' }
        query_type: { $ref: '#/components/schemas/DnsQueryType' }
        response_code: { $ref: '#/components/schemas/DnsResponseCode' }
        transport: { $ref: '#/components/schemas/DnsTransport' }
        direction: { $ref: '#/components/schemas/DnsDirection' }
    FileActivitySemanticSummary:
      type: object
      additionalProperties: false
      required: [operation, process_command, path]
      properties:
        operation: { $ref: '#/components/schemas/FileActivityOperation' }
        process_command: { type: string, minLength: 1 }
        path: { type: string, minLength: 1, description: Path reported by the process syscall. }
        new_path:
          {
            type: string,
            minLength: 1,
            description: Rename destination path reported by the process syscall.,
          }
        replaced:
          {
            type: [boolean, 'null'],
            description: Whether rename replaced an existing destination; null means unknown.,
          }
...
    ProcessExitSemanticSummary:
      type: object
      additionalProperties: true
      required: [evidence_source, termination]
      properties:
        evidence_source: { type: string, const: kernel }
        termination: { $ref: '#/components/schemas/ProcessTermination' }
    ContainerTerminationSemanticSummary:
      type: object
      additionalProperties: true
      required: [evidence_source, container_name, reason, exit_code]
      properties:
        evidence_source: { type: string, const: kubernetes }
        container_name: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1 }
        exit_code: { type: integer, minimum: 0 }
    ContainerRestartSemanticSummary:
      type: object
      additionalProperties: true
      required: [evidence_source, container_name]
      properties:
        evidence_source: { type: string, const: kubernetes }
        container_name: { type: string, minLength: 1 }
    RestartLoopSemanticSummary:
      type: object
      additionalProperties: false
      required: [evidence_source, projection_version, threshold, window_started_at, window_ended_at, observed_restart_count, container_name]
      properties:
        evidence_source: { type: string, const: derived }
        projection_version: { type: integer, minimum: 1 }
        threshold: { type: integer, minimum: 1 }
        window_started_at: { $ref: '#/components/schemas/Timestamp' }
        window_ended_at: { $ref: '#/components/schemas/Timestamp' }
        observed_restart_count: { type: integer, minimum: 1 }
        container_name: { type: string, minLength: 1 }
        latest_termination: { $ref: '#/components/schemas/KubernetesTermination' }
        latest_waiting_reason: { type: [string, 'null'], maxLength: 256 }
...
    DnsQueryType: { type: string, enum: [A, AAAA] }
    DnsTransport: { type: string, enum: [udp, tcp] }
    DnsDirection: { type: string, enum: [egress, ingress] }
...
    DnsContext:
      type: object
      additionalProperties: false
      required: [names, observed_at, expires_at, confidence, ambiguous]
      properties:
        names:
          {
            type: array,
            minItems: 1,
            maxItems: 8,
            uniqueItems: true,
            items: { $ref: '#/components/schemas/DnsName' },
          }
        observed_at: { $ref: '#/components/schemas/Timestamp' }
        expires_at: { $ref: '#/components/schemas/Timestamp' }
        confidence: { $ref: '#/components/schemas/DnsConfidence' }
        ambiguous: { type: boolean }
...
    Release:
      type: object
      additionalProperties: false
      required: [id, project_id, application_id, version, description, deployed_at, created_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        application_id: { $ref: '#/components/schemas/Uuid' }
        version: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, 'null'], maxLength: 2000 }
        deployed_at: { $ref: '#/components/schemas/Timestamp' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
    ReleasePage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Release' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
    CreateRelease:
      type: object
      additionalProperties: false
      required: [version, deployed_at]
      properties:
        version: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, 'null'], maxLength: 2000 }
        deployed_at: { $ref: '#/components/schemas/Timestamp' }
    RuntimeDiffEntry:
      type: object
      additionalProperties: false
      required:
        [
          group_id,
          classification,
          event_kind,
          semantic_summary,
          baseline_occurrence_count,
          baseline_first_seen_at,
          baseline_last_seen_at,
          target_occurrence_count,
          target_first_seen_at,
          target_last_seen_at,
        ]
      properties:
        group_id: { $ref: '#/components/schemas/Uuid' }
        classification: { type: string, enum: [new, disappeared, unchanged] }
        event_kind: { type: string }
        semantic_summary: { $ref: '#/components/schemas/RuntimeEventSemanticSummary' }
        baseline_occurrence_count: { type: [integer, 'null'], format: int64 }
        baseline_first_seen_at: { $ref: '#/components/schemas/NullableTimestamp' }
        baseline_last_seen_at: { $ref: '#/components/schemas/NullableTimestamp' }
        target_occurrence_count: { type: [integer, 'null'], format: int64 }
        target_first_seen_at: { $ref: '#/components/schemas/NullableTimestamp' }
        target_last_seen_at: { $ref: '#/components/schemas/NullableTimestamp' }
    RuntimeDiff:
      type: object
      additionalProperties: false
      required: [baseline, target, items, next_cursor]
      properties:
        baseline: { oneOf: [{ $ref: '#/components/schemas/Release' }, { type: 'null' }] }
        target: { $ref: '#/components/schemas/Release' }
        items: { type: array, items: { $ref: '#/components/schemas/RuntimeDiffEntry' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
    RuntimeDiffClassificationCount:
      type: object
      additionalProperties: false
      required: [classification, item_count]
      properties:
        classification: { type: string, enum: [new, disappeared, unchanged] }
        item_count: { type: integer, format: int64, minimum: 0 }
    RuntimeDiffChangeEntry:
      type: object
      additionalProperties: false
      required:
        [
          group_id,
          classification,
          event_kind,
          semantic_summary,
          baseline_occurrence_count,
          target_occurrence_count,
          occurrence_delta,
        ]
      properties:
        group_id: { $ref: '#/components/schemas/Uuid' }
        classification: { type: string, enum: [new, disappeared, unchanged] }
        event_kind: { type: string }
        semantic_summary: { $ref: '#/components/schemas/RuntimeEventSemanticSummary' }
        baseline_occurrence_count: { type: integer, format: int64, minimum: 0 }
        target_occurrence_count: { type: integer, format: int64, minimum: 0 }
        occurrence_delta: { type: integer, format: int64 }
    RuntimeDiffSummary:
      type: object
      additionalProperties: false
      required: [baseline, target, total_item_count, classifications, largest_changes]
      properties:
        baseline: { oneOf: [{ $ref: '#/components/schemas/Release' }, { type: 'null' }] }
        target: { $ref: '#/components/schemas/Release' }
        total_item_count: { type: integer, format: int64, minimum: 0 }
        classifications:
          {
            type: array,
            maxItems: 3,
            items: { $ref: '#/components/schemas/RuntimeDiffClassificationCount' },
          }
        largest_changes:
          {
            type: array,
            maxItems: 10,
            items: { $ref: '#/components/schemas/RuntimeDiffChangeEntry' },
          }
...
    Uuid: { type: string, format: uuid }
    NullableUuid: { type: [string, 'null'], format: uuid }
...
    Timestamp: { type: string, format: date-time, example: '2026-08-17T12:00:00Z' }
    NullableTimestamp: { type: [string, 'null'], format: date-time }
  requestBodies:
    CreateRelease:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/CreateRelease' } } }
...
  responses:
...
    Release:
      {
        description: Release,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/Release' } } },
      }
    ReleasePage:
      {
        description: Release page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/ReleasePage' } } },
      }
    RuntimeDiff:
      {
        description: Runtime behavior diff,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/RuntimeDiff' } } },
      }
    RuntimeDiffSummary:
      {
        description: Complete runtime comparison aggregate,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/RuntimeDiffSummary' } } },
      }
...
    Error:
      description: Correlated API error
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }